\name{tempcont_data}
\alias{tempcont_data}
\docType{data}
\title{
Dummy data for \code{temp.cont} examples
}
\description{
Data from field work in the Middle-earth.
}
\usage{data("tempcont_data")}
\format{
  A data frame with 400 observations on the following 10 variables.
  \describe{
    \item{\code{cave}}{a factor vector}
    \item{\code{year}}{a numeric vector with a temporal variable}
    \item{\code{trolls}}{a numeric vector}
    \item{\code{bushes}}{a numeric vector}
    \item{\code{temperature}}{a numeric vector}
    \item{\code{rainfall}}{a numeric vector}
    \item{\code{drought}}{a numeric vector}
    \item{\code{gremlins}}{a numeric vector}
    \item{\code{orcs}}{a numeric vector}
    \item{\code{warlocks}}{a numeric vector}
  }
}
\examples{
data(tempcont_data)
str(tempcont_data)
}
\keyword{datasets}
