\name{temp.cont}
\alias{temp.cont}
\title{Temporal contribution}
\usage{
  temp.cont(model, resp, driver, random, dataframe)
}
\arguments{
  \item{model}{}
  \item{resp}{}
  \item{driver}{}
  \item{random}{}
  \item{dataframe}{}
}
\description{
Prints 'temp.cont, world!'.
}
\value{}
\examples{
temp.cont()
}
